requirements.txt
ipykernel
pandas
pyYAML
python-box
types-PyYAML
psycopg2
python-dotenv
scikit-learn
joblib
mlflow